+++ tide/core/data/Image.h
@@ -80,9 +80,17 @@ public:
         return tex.width() * tex.height() * bpp;
     }
 
+    /** @return the row order of the image data. */
+    virtual deflect::RowOrder getRowOrder() const
+    {
+        return deflect::RowOrder::top_down;
+    }
+
     /** @return the format of the image. */
     virtual TextureFormat getFormat() const = 0;
 
+    /** @return the color space of the image. */
+    virtual ColorSpace getColorSpace() const { return ColorSpace::undefined; }
     /** @return the OpenGL pixel format of the image data. */
     virtual uint getGLPixelFormat() const = 0;